Update the SME operation sample
